+++ tests/testthat/test-parse-glycam-iupac.R
@@ -41,9 +41,9 @@ test_that("GlyCAM IUPAC preserves unusual configurations", {
     as.character(parsed),
     c(
       natural_fucose = "Fuc(a1-",
-      d_fucose = "DFuc3S(a1-",
-      l_gulose = "LGul(b1-",
-      d_fucofuranose = "DFucf(a1-"
+      d_fucose = "D-Fuc3S(a1-",
+      l_gulose = "L-Gul(b1-",
+      d_fucofuranose = "D-Fucf(a1-"
     )
   )
 })
@@ -172,7 +172,7 @@ test_that("GlyCAM IUPAC parses checklist monosaccharide names", {
       Dig = "Dig(b1-",
       Col = "Col(b1-",
       Ara = "Ara(b1-",
-      Lyx = "LLyx(b1-",
+      Lyx = "L-Lyx(b1-",
       Xyl = "Xyl(b1-",
       Rib = "Rib(b1-",
       Neu5Ac = "Neu5Ac(b2-",
@@ -190,7 +190,7 @@ test_that("GlyCAM IUPAC parses checklist monosaccharide names", {
       MurNAc = "MurNAc(b1-",
       MurNGc = "MurNGc(b1-",
       Mur = "Mur(b1-",
-      Api = "DApif(b1-",
+      Api = "D-Apif(b1-",
       Fru = "Fru(b2-",
       Tag = "Tag(b2-",
       Sor = "Sor(b2-",